Aggregate complex build configuration
