+++ packages/tests/src/plugins/telemetry/index.test.ts
@@ -7,7 +7,8 @@ import { getMetrics } from '@dd/telemetry-plugins/common/aggregator';
 import type { MetricToSend } from '@dd/telemetry-plugins/types';
 import type { Bundler, CleanupFn } from '@dd/tests/helpers/runBundlers';
 import { BUNDLERS, runBundlers } from '@dd/tests/helpers/runBundlers';
-import path from 'path';
+
+import { getComplexBuildOverrides } from '../../helpers/mocks';
 
 // Used to intercept metrics.
 jest.mock('@dd/telemetry-plugins/common/aggregator', () => {
@@ -19,33 +20,6 @@ jest.mock('@dd/telemetry-plugins/common/aggregator', () => {
 });
 
 const getMetricsMocked = jest.mocked(getMetrics);
-
-const entries = {
-    app1: '@dd/tests/fixtures/project/main1.js',
-    app2: '@dd/tests/fixtures/project/main2.js',
-};
-
-const bundlerOverrides = {
-    rollup: {
-        input: entries,
-    },
-    vite: {
-        input: entries,
-    },
-    esbuild: {
-        entryPoints: entries,
-    },
-    webpack5: { entry: entries },
-    webpack4: {
-        // Webpack 4 doesn't support pnp.
-        entry: Object.fromEntries(
-            Object.entries(entries).map(([name, filepath]) => [
-                name,
-                `./${path.relative(process.cwd(), require.resolve(filepath))}`,
-            ]),
-        ),
-    },
-};
 
 const getGetMetricsImplem: (metrics: Record<string, MetricToSend[]>) => typeof getMetrics =
     (metrics) => (context, options, report) => {
@@ -118,7 +92,11 @@ describe('Telemetry Universal Plugin', () => {
                 };
                 // This one is called at initialization, with the initial context.
                 getMetricsMocked.mockImplementation(getGetMetricsImplem(metrics));
-                cleanup = await runBundlers(pluginConfig, bundlerOverrides, activeBundlers);
+                cleanup = await runBundlers(
+                    pluginConfig,
+                    getComplexBuildOverrides(),
+                    activeBundlers,
+                );
             });
 
             afterAll(async () => {
